+++ src/main/java/de/geheimagentnr1/manyideas_core/elements/blocks/dye_crafting_table/DyeCraftingTableScreen.java
@@ -31,14 +31,14 @@ public class DyeCraftingTableScreen extends AbstractContainerScreen<DyeCraftingT
 	
 	private void initData() {
 		
-		titleLabelX = 38;
+		titleLabelX = 29;
 	}
 	
 	@Override
 	public void render( @NotNull GuiGraphics guiGraphics, int mouseX, int mouseY, float partialTick ) {
 		
-		renderBackground( guiGraphics, mouseX, mouseY, partialTick );
 		super.render( guiGraphics, mouseX, mouseY, partialTick );
+		renderTooltip( guiGraphics, mouseX, mouseY );
 	}
 	
 	@Override